requirements.txt
django
djangorestframework
pika
pydantic
retry
